% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{color}
\alias{color}
\title{color}
\usage{
color(input_path)
}
\arguments{
\item{input_path}{string, path for the input png file}
}
\value{
a png file at the same path as input_path
}
\description{
color
}
\examples{
#' color("../img/test.png")

}
