% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenscale.R
\name{greenscale}
\alias{greenscale}
\title{Title}
\usage{
greenscale(input_path)
}
\arguments{
\item{input_path}{Path of the input image}
}
\value{
a png file at the same path as input_path
}
\description{
Title
}
