% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror.R
\name{mirror}
\alias{mirror}
\title{Flip}
\usage{
mirror(input_path)
}
\arguments{
\item{input_path}{string, path for the input png file}
}
\value{
a png file at the same path as input_path
}
\description{
Flip
}
\examples{
#' mirror("../img/random.png")

}
