% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpen.R
\name{sharpen}
\alias{sharpen}
\title{Sharpened}
\usage{
sharpen(input_path)
}
\arguments{
\item{input_path}{string, path for the input png file}
}
\value{
a png file at the same path as input_path
}
\description{
Sharpened
}
\examples{
#' sharpen("../img/test.png")

}
