% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror.R
\name{mirror}
\alias{mirror}
\title{Returns a mirrored image}
\usage{
mirror(input_path)
}
\arguments{
\item{input_path}{string, path for the input png file}
}
\value{
an image file at the same path as input_path and returns the output path
}
\description{
Returns a mirrored image
}
\examples{
#' mirror("test_img/test_original.jpg")
}
